refactor PromptTemplate class and sanitize_parameters function
