+++ openllm-core/src/openllm_core/__init__.py
@@ -42,3 +42,5 @@ from .config import MPTConfig as MPTConfig
 from .config import OPTConfig as OPTConfig
 from .config import StableLMConfig as StableLMConfig
 from .config import StarCoderConfig as StarCoderConfig
+from .prompt import PromptTemplate as PromptTemplate
+from .prompt import process_prompt as process_prompt
