+++ openllm-core/src/openllm_core/config/configuration_dolly_v2.py
@@ -3,7 +3,8 @@ import typing as t
 
 import openllm_core
 
-from openllm_core.prompts._prompt import process_prompt
+from openllm_core.prompt import PromptTemplate
+from openllm_core.prompt import process_prompt
 from openllm_core.utils import dantic
 
 if t.TYPE_CHECKING:
@@ -91,7 +92,8 @@ class DollyV2Config(openllm_core.LLMConfig):
 
   def sanitize_parameters(self,
                           prompt: str,
-                          prompt_template: str | None = None,
+                          prompt_template: PromptTemplate | str | None = None,
+                          system_message: str | None = None,
                           max_new_tokens: int | None = None,
                           temperature: float | None = None,
                           top_k: int | None = None,
