+++ openllm-core/src/openllm_core/config/configuration_flan_t5.py
@@ -3,7 +3,8 @@ import typing as t
 
 import openllm_core
 
-from openllm_core.prompts._prompt import process_prompt
+from openllm_core.prompt import PromptTemplate
+from openllm_core.prompt import process_prompt
 
 START_FLAN_T5_COMMAND_DOCSTRING = '''\
 Run a LLMServer for FLAN-T5 model.
@@ -56,7 +57,8 @@ class FlanT5Config(openllm_core.LLMConfig):
 
   def sanitize_parameters(self,
                           prompt: str,
-                          prompt_template: str | None = None,
+                          prompt_template: PromptTemplate | str | None = None,
+                          system_message: str | None = None,
                           max_new_tokens: int | None = None,
                           temperature: float | None = None,
                           top_k: int | None = None,
