+++ openllm-core/src/openllm_core/config/configuration_gpt_neox.py
@@ -3,7 +3,8 @@ import typing as t
 
 import openllm_core
 
-from openllm_core.prompts._prompt import process_prompt
+from openllm_core.prompt import PromptTemplate
+from openllm_core.prompt import process_prompt
 from openllm_core.utils import dantic
 
 START_GPT_NEOX_COMMAND_DOCSTRING = '''\
@@ -58,7 +59,8 @@ class GPTNeoXConfig(openllm_core.LLMConfig):
 
   def sanitize_parameters(self,
                           prompt: str,
-                          prompt_template: str | None = None,
+                          prompt_template: PromptTemplate | str | None = None,
+                          system_message: str | None = None,
                           temperature: float | None = None,
                           max_new_tokens: int | None = None,
                           use_default_prompt_template: bool = True,
