+++ openllm-core/src/openllm_core/config/configuration_mpt.py
@@ -3,7 +3,8 @@ import typing as t
 
 import openllm_core
 
-from openllm_core.prompts._prompt import process_prompt
+from openllm_core.prompt import PromptTemplate
+from openllm_core.prompt import process_prompt
 from openllm_core.utils import dantic
 
 MPTPromptType = t.Literal['default', 'instruct', 'chat', 'storywriter']
@@ -86,7 +87,8 @@ class MPTConfig(openllm_core.LLMConfig):
 
   def sanitize_parameters(self,
                           prompt: str,
-                          prompt_template: str | None = None,
+                          prompt_template: PromptTemplate | str | None = None,
+                          system_message: str | None = None,
                           max_new_tokens: int | None = None,
                           temperature: float | None = None,
                           top_p: float | None = None,
