+++ openllm-core/src/openllm_core/config/configuration_opt.py
@@ -3,11 +3,11 @@ import typing as t
 
 import openllm_core
 
-from openllm_core.prompts._prompt import process_prompt
+from openllm_core.prompt import process_prompt
 from openllm_core.utils import dantic
 
 if t.TYPE_CHECKING:
-  from openllm_core.prompts.prompt_template import PromptTemplate
+  from openllm_core.prompt.prompt_template import PromptTemplate
 
 START_OPT_COMMAND_DOCSTRING = '''\
 Run a LLMServer for OPT model.
@@ -67,7 +67,8 @@ class OPTConfig(openllm_core.LLMConfig):
 
   def sanitize_parameters(self,
                           prompt: str,
-                          prompt_template: PromptTemplate | None = None,
+                          prompt_template: PromptTemplate | str | None = None,
+                          system_message: str | None = None,
                           max_new_tokens: int | None = None,
                           temperature: float | None = None,
                           top_k: int | None = None,
