+++ openllm-core/src/openllm_core/config/configuration_stablelm.py
@@ -3,7 +3,8 @@ import typing as t
 
 import openllm_core
 
-from openllm_core.prompts._prompt import process_prompt
+from openllm_core.prompt import PromptTemplate
+from openllm_core.prompt import process_prompt
 
 START_STABLELM_COMMAND_DOCSTRING = '''\
 Run a LLMServer for StableLM model.
@@ -62,7 +63,8 @@ class StableLMConfig(openllm_core.LLMConfig):
 
   def sanitize_parameters(self,
                           prompt: str,
-                          prompt_template: str | None = None,
+                          prompt_template: PromptTemplate | str | None = None,
+                          system_message: str | None = None,
                           temperature: float | None = None,
                           max_new_tokens: int | None = None,
                           top_k: int | None = None,
