+++ openllm-core/src/openllm_core/config/configuration_starcoder.py
@@ -3,6 +3,9 @@ import typing as t
 
 import openllm_core
 
+if t.TYPE_CHECKING:
+  from openllm_core.prompt import PromptTemplate
+
 START_STARCODER_COMMAND_DOCSTRING = '''\
 Run a LLMServer for StarCoder model.
 
@@ -54,7 +57,8 @@ class StarCoderConfig(openllm_core.LLMConfig):
 
   def sanitize_parameters(self,
                           prompt: str,
-                          prompt_template: str | None = None,
+                          prompt_template: PromptTemplate | str | None = None,
+                          system_message: str | None = None,
                           temperature: float | None = None,
                           top_p: float | None = None,
                           max_new_tokens: int | None = None,
