+++ openllm-core/src/openllm_core/prompt/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import annotations
+
+from .prompt_template import PromptTemplate as PromptTemplate
+from .prompt_template import process_prompt as process_prompt
