--- openllm-core/src/openllm_core/prompts/__init__.py
@@ -1,1 +1,0 @@
-from __future__ import annotations
